configure webpack
