+++ package.json
@@ -1,11 +1,12 @@
 {
   "name": "Todo-list-js",
   "version": "1.0.0",
-  "private": true,
   "description": "This is a JavaScript To-Do list app where users can add projects and there corresponding task lists they like to work on.",
-  "main": "index.js",
+  "private": true,
   "scripts": {
-    "start": "webpack --config webpack.common.js"
+    "start": "webpack --config webpack.dev.js",
+    "build": "webpack --config webpack.prod.js",
+    "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
     "type": "git",
@@ -19,16 +20,26 @@
   },
   "homepage": "https://github.com/tGodson/Todo-list-js#readme",
   "devDependencies": {
+    "babel-eslint": "^10.1.0",
     "bootstrap": "^4.5.3",
+    "clean-webpack-plugin": "^3.0.0",
     "css-loader": "^5.0.0",
+    "eslint": "^7.12.1",
+    "eslint-config-airbnb-base": "^14.2.0",
     "file-loader": "^6.2.0",
     "html-loader": "^1.3.2",
     "html-webpack-plugin": "^4.5.0",
     "node-sass": "^4.14.1",
     "sass-loader": "^10.0.4",
     "style-loader": "^2.0.0",
+    "stylelint": "^13.7.2",
+    "stylelint-config-standard": "^20.0.0",
+    "stylelint-csstree-validator": "^1.9.0",
+    "stylelint-scss": "^3.18.0",
     "webpack": "^5.2.0",
-    "webpack-cli": "^4.1.0"
+    "webpack-cli": "^4.1.0",
+    "webpack-dev-server": "^3.11.0",
+    "webpack-merge": "^5.2.0"
   },
   "dependencies": {
     "@fortawesome/fontawesome-free": "^5.15.1",
