+++ src/template.html
@@ -81,9 +81,6 @@
     </div>
   </div>
 </div>
-<!-- <script src="index.js"></script>
-<script src="logic.js"></script> -->
-<!-- <script src="../dist/main.js"></script> -->
 </body>
   
 </html>
